pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
vfab = ["config/*.yaml"]

[project]
name = "vfab"
version = "0.9.0"
description = "vfab: FSM plotter manager (headless-first) with vpype + per-session recording"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.12.3",
    "pydantic>=2.9",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.36",
    "alembic>=1.13.2",
    "rich>=13.0.0",
    "jinja2>=3.1.0",
    "psutil>=5.9.0",
    "click>=8.0.0",
    "platformdirs>=4.0.0",
    "defusedxml>=0.7.1",
    "fastapi>=0.104.0",
    "websockets>=12.0",
    "uvicorn[standard]>=0.24.0",
]

[project.optional-dependencies]
vpype = ["vpype>=1.14.0"]
# NOTE: EMS API is distributed as a zip; import name is `pyaxidraw`
# Package metadata name is `axicli` but provides `pyaxidraw` module
axidraw = ["axicli @ https://cdn.evilmadscientist.com/dl/ad/public/AxiDraw_API.zip"]
dev = ["pytest>=8.3.2", "pre-commit>=3.8.0", "ruff>=0.6.8", "black>=24.10.0"]

[project.scripts]
vfab = "vfab.cli:main"

[dependency-groups]
dev = ["pytest>=8.3.2", "pre-commit>=3.8.0", "ruff>=0.6.8", "black>=24.10.0"]
vpype = ["vpype>=1.14.0"]

[tool.uv]
default-groups = ["dev"]   # keep axidraw opt-in

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B110", "B112"]
